# application/services/custom_services/generalledger/src/models/daomodels/generalledger.ts
import * as mongoose from 'mongoose';

const Schema = mongoose.Schema;

export const generalledgerSchema = new Schema({
   created_date: { type: Date, default: Date.now },
   created_by: { type: String },
   last_modified_by: { type: String },
   last_modified_date: { type: Date, default: Date.now },
   name: String
})

const generalledgerModel = mongoose.model('generalledger', generalledgerSchema, 'generalledger');
export default generalledgerModel;